fix: remove empty profiles from lock file generations
