+++ src/main/java/se/vandmo/dependencylock/maven/json/WithJsonHelper.java
@@ -110,6 +110,7 @@ class WithJsonHelper {
   JsonNode buildProfilesJson(Stream<ProfileEntry> profiles, JsonNodeFactory jsonNodeFactory) {
     ArrayNode result = jsonNodeFactory.arrayNode();
     profiles
+        .filter(entry -> !entry.isEmpty())
         .sorted(Comparator.comparing(entry -> entry.getProfile().getId()))
         .forEach(
             profile -> {
