+++ src/main/java/se/vandmo/dependencylock/maven/pom/LockFilePom.java
@@ -1,24 +1,18 @@
 package se.vandmo.dependencylock.maven.pom;
 
-import static freemarker.template.TemplateExceptionHandler.RETHROW_HANDLER;
 import static java.lang.String.format;
 import static java.util.Locale.ROOT;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
 
 import freemarker.template.Configuration;
-import freemarker.template.DefaultObjectWrapperBuilder;
-import freemarker.template.ObjectWrapper;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
-import freemarker.template.Version;
 import java.io.File;
 import java.io.IOException;
 import java.io.Writer;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import org.apache.maven.plugin.MojoExecutionException;
@@ -34,9 +28,7 @@ import se.vandmo.dependencylock.maven.PomMinimums;
 import se.vandmo.dependencylock.maven.ProfileEntry;
 import se.vandmo.dependencylock.maven.ProfiledDependencies;
 
-public final class LockFilePom implements Lockfile {
-
-  private static final Version VERSION = Configuration.VERSION_2_3_31;
+public final class LockFilePom extends WithLockFilePom implements Lockfile {
 
   private final LockFileAccessor dependenciesLockFile;
   private final PomMinimums pomMinimums;
@@ -75,48 +67,12 @@ public final class LockFilePom implements Lockfile {
     }
   }
 
-  private static Map<String, Object> makeDataModel(
-      PomMinimums pomMinimums, LockedProject lockedProject) {
-    Map<String, Object> dataModel = new HashMap<>();
-    dataModel.put("pom", pomMinimums);
-    dataModel.put("dependencies", lockedProject.dependencies.getSharedDependencies());
+  private Map<String, Object> makeDataModel(PomMinimums pomMinimums, LockedProject lockedProject) {
+    Map<String, Object> dataModel = super.makeDataModel(pomMinimums, lockedProject.dependencies);
     lockedProject.parents.ifPresent(parents -> dataModel.put("parents", parents));
     lockedProject.plugins.ifPresent(plugins -> dataModel.put("plugins", plugins));
     lockedProject.extensions.ifPresent(extensions -> dataModel.put("extensions", extensions));
-    dataModel.put(
-        "profiles",
-        lockedProject
-            .dependencies
-            .profileEntries()
-            .sorted(Comparator.comparing(a -> a.getProfile().getId()))
-            .map(
-                entry -> {
-                  Map<String, Object> result = new HashMap<>();
-                  result.put("id", entry.getProfile().getId());
-                  result.put("activation", entry.getProfile().getActivation());
-                  result.put("dependencies", entry.getDependencies());
-                  return result;
-                })
-            .collect(toList()));
     return dataModel;
-  }
-
-  private static Configuration createConfiguration() {
-    Configuration cfg = new Configuration(VERSION);
-    cfg.setClassForTemplateLoading(LockFilePom.class, "");
-    cfg.setDefaultEncoding("UTF-8");
-    cfg.setTemplateExceptionHandler(RETHROW_HANDLER);
-    cfg.setLogTemplateExceptions(false);
-    cfg.setFallbackOnNullLoopVariable(false);
-    cfg.setObjectWrapper(getObjectWrapper());
-    return cfg;
-  }
-
-  private static ObjectWrapper getObjectWrapper() {
-    DefaultObjectWrapperBuilder builder = new DefaultObjectWrapperBuilder(VERSION);
-    builder.setExposeFields(true);
-    builder.setIterableSupport(true);
-    return builder.build();
   }
 
   @Override
